Refactor main.go and addPassword.go
