+++ src/main.go
@@ -1,26 +1,23 @@
 package main
 
 import (
-	"fmt"
 	"log"
 	"os"
 
-	"github.com/playjeri/passutasku/src/manager"
+	m "github.com/playjeri/passutasku/src/manager"
 	"github.com/playjeri/passutasku/src/utils"
 
 	tea "github.com/charmbracelet/bubbletea"
-	m "github.com/playjeri/passutasku/src/manager/models"
 )
 
 func main() {
 	defer utils.ClearTerminal()
 	utils.ClearTerminal()
 
-	file := manager.OpenLogFile()
+	file := m.OpenLogFile()
 	log.SetOutput(file)
-	defer file.Close()
 
-	data := manager.LoadFile()
+	data := m.LoadFile()
 	MainModel := m.MainModel{
 		State: m.ShowPasswords,
 		ShowPasswordsModel: m.ShowPasswordsModel{
@@ -28,18 +25,19 @@ func main() {
 		},
 		AddPasswordModel: m.NewModel(),
 	}
-	defer log.Println(MainModel.ShowPasswordsModel.Passwords)
-
-	defer manager.SaveFile(MainModel.ShowPasswordsModel.Passwords)
 
 	if len(MainModel.ShowPasswordsModel.Passwords) == 0 {
-		log.Println("No passwords found in the file")
-		MainModel.ShowPasswordsModel = m.TestModel()
+		os.Exit(1)
+		MainModel.ShowPasswordsModel.Passwords = []m.PasswordEntry{
+			{Service: "Twitter", Username: "user1", Password: "password1"},
+			{Service: "Facebook", Username: "user2", Password: "password2"},
+			{Service: "Instagram", Username: "user3", Password: "password3"},
+		}
 	}
 
 	p := tea.NewProgram(MainModel)
 	if _, err := p.Run(); err != nil {
-		fmt.Printf("Alas, there's been an error: %v", err)
 		os.Exit(1)
 	}
+	defer file.Close()
 }
